refactor EditorView and introduce tab pane
This moves the run button up to a context menu at the top. This is also the menu where future file handling will go, so there's no need for Run to take up more real estate at the bottom of the screen. This also refactors the main EditorView so that updates are done by the controller instead.