creates basic version of IfExpression and LetBinding classes
This will allow a more rich form of understanding these types of expressions once they are parsed. The form of an IfExpression or LetBinding doesn't need to be duplicated in the Parser and in the Executor. Once the parser knows something is a let binding or an if expression, it should encode that knowledge into the type system itself.