Refactor keywords (#4)
* creates basic version of IfExpression and LetBinding classes

This will allow a more rich form of understanding these types of expressions once they are parsed. The form of an IfExpression or LetBinding doesn't need to be duplicated in the Parser and in the Executor. Once the parser knows something is a let binding or an if expression, it should encode that knowledge into the type system itself.

* enriches the type of Expression when found in Parser

When we encounter a let binding or an if expression, return the fancier type we just made. This is 100% compatible with the current code as they both inherit from Expression.

* change Executor to defer to IfBinding and LetBinding execute

They know how to do it best themselves.

* accept some new warnings

* move IfExpression and LetBinding into new files

* create Symbol class

* create constructors for ExpressionPart to replace helper functions

* rename type field of Data to dataType

I want this to become an ExpressionPart, but the "type" field will conflict.

* make Function an Expression

* make Data an ExpressionPartType

* Removed simple ExpressionPart accessory functions

* remove the symbolPart and stringPart

* remove createData accessory function

* extract Symbol to its own file

* create FunctionDefinition object

* change Executor to use FunctionDefinition

* move types into expressions package

* small cleanup

* change to server I forgot to checkin

* change Scope to accept only Symbols

* makes Expression an ExpressionPart

* introduce Keyword type

* make ExpressionPart abstract

* cleanup ExpressionPartType

* move Keyword into new file

* move ExpressionPart into expressions package