change built-in handling to scope registration
This uses reflection to build a list of all final subclasses of
BuiltInFunction. At the creation of any new Scope, these are then
instantiated and populated by name in the Scope.

This lets Executor not care about built-in functions at all and we not
longer have to maintain a separate list of function names to classes.